Projekte/Materialverwaltung/requirements.txt
anyio==4.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.1
coverage==7.9.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
fastjsonschema==2.20.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
iniconfig==1.1.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mistune==3.1.2
mypy==1.14.1
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.3.1
overrides==7.4.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.8.0
pip==25.1
platformdirs==4.3.7
pluggy==1.5.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
Send2Trash==1.8.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.39
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==2.3.3
